+++ lib/papa/task/common/finish.rb
@@ -14,22 +14,13 @@ module Papa
         def run
           @build_branch ||= "#{@build_type}/#{@version}"
 
+          check_if_build_branch_exists
+
           success = true
           @success_branches = []
 
           @base_branches.each do |branch|
-            queue = [
-              Command::Git::Fetch.new('origin'),
-              Command::Git::Checkout.new(@build_branch),
-              Command::Git::Checkout.new(branch),
-              Command::Git::Merge.new(@build_branch),
-              Command::Git::Push.new('origin', branch)
-            ]
-            if @tag_name && branch == 'master'
-              queue << Command::Git::Tag.new(@tag_name)
-              queue << Command::Git::TagPush.new('origin', @tag_name)
-            end
-            runner = Runner.new(queue)
+            runner = Runner.new(queue(branch))
 
             if runner.run
               @success_branches << branch
@@ -38,15 +29,40 @@ module Papa
             end
           end
 
-          success_message if !@success_branches.empty?
-
-          if !success
+          if success
+            success_message
+          else
             failure_message
             exit 1
           end
         end
 
         private
+
+        def check_if_build_branch_exists
+          queue = [
+            Command::Git::Fetch.new('origin'),
+            Command::Git::Checkout.new(@build_branch)
+          ]
+          runner = Runner.new(queue)
+          return if runner.run
+          Helper::Output.failure 'Build branch does not exist.'
+          exit 1
+        end
+
+        def queue(branch)
+          queue = [
+            Command::Git::Checkout.new(@build_branch),
+            Command::Git::Checkout.new(branch),
+            Command::Git::Merge.new(@build_branch),
+            Command::Git::Push.new('origin', branch)
+          ]
+          if @tag_name && branch == 'master'
+            queue << Command::Git::Tag.new(@tag_name)
+            queue << Command::Git::TagPush.new('origin', @tag_name)
+          end
+          queue
+        end
 
         def success_message
           Helper::Output.success "Successfully merged #{@build_branch} to these branches:\n"